+++ l10n/da.js
@@ -2,9 +2,12 @@ OC.L10N.register(
     "integration_dropbox",
     {
     "Dropbox" : "Dropbox",
+    "Bad HTTP method" : "Dårlig HTTP metode",
+    "Bad credentials" : "Forkerte legitimationsoplysninger",
     "Connected accounts" : "Forbundne konti",
-    "App key" : "App-nøgle",
-    "App secret" : "App-hemmelighed",
-    "Authentication" : "Godkendelse"
+    "App key" : "App nøgle",
+    "App secret" : "App hemmelighed",
+    "Authentication" : "Godkendelse",
+    "Connected as {user} ({email})" : "Forbundet som {user} ({email})"
 },
 "nplurals=2; plural=(n != 1);");